% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TTS2016R-package.R
\docType{data}
\name{ggh_cma}
\alias{ggh_cma}
\title{census metropolitan area (CMA)/census agglomeration (CA) boundaries in the Greater Golden Horseshoe (GGH) area and associated data.}
\format{
A simple feature class (sf) polygon object containing 16 rows and 4 variables; each row represents a unique census metropolitan area (CMA) of census agglomeration (CA) with associated geo-referenced geometry. CMA and CA consist of one or more adjacent municipalities around a core. The CMA/CA boundaries are defined for the purpose of the Canadian Census based on population. As such, the core of each CMA must have a population of at least 50,000 and the entire CMA must have a population of at least 100,000. CA are smaller in population, thir course must have a population of at least 10,000.
\describe{
\item{CMAUID}{A unique identifier for each census metropolitan area (CMA)/census agglomeration (CA).}
\item{CMANAME}{The name of the census metropolitan area (CMA)/census agglomeration (CA).}
\item{CMATYPE}{A one character field indicating whether the unit is a census metropolitan area, a tracted census agglomeration or a non-tracted census agglomeration. "B" is a CMA and "D" is a CA with no census tracts, and "K" is a CA with census tracts.}
\item{geometry}{The sfc polygon geometry (boundaries).}
}
}
\source{
"The 2016 census metropolitan areas and census agglomeration boundary files as created by the Canadian Census available \href{https://www12.statcan.gc.ca/census-recensement/2011/geo/bound-limit/bound-limit-2016-eng.cfm}{here} accessed October 21st 2022 (Boundary Files, 2016 Census. Statistics Canada Catalogue no. 92-160-X.). All variable definitions are based on the definitions included in the census year 2016 boundary file reference guide (Boundary Files, Reference Guide, Second edition, 2016 Census. Statistics Canada Catalogue no. 92-160-G.).
}
\usage{
data(ggh_cma)
}
\description{
This object contains the census metropolitan area (CMA)/census agglomeration (CA) for the spatial region  associated with the 2016 Transportation Tomorrow Survey (TTS) in the Greater Golden Horseshoe (GGH) area.
}
\keyword{boundary}
\keyword{districts}
\keyword{municipalities}
\keyword{planning}
\keyword{regions}
