% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TTS2016R-package.R
\docType{data}
\name{ggh_pd}
\alias{ggh_pd}
\title{Planning boundaries in the Greater Golden Horseshoe (GGH) area and associated data.}
\format{
A simple feature class (sf) polygon object containing 105 rows and 5 variables; each row represents a unique planning region with associated geo-referenced geometry.
\describe{
\item{MUN}{The name of the municipality and/or planning boundary.}
\item{PD}{A unique ID.}
\item{REGION}{A unique ID corresponding with the greater region that each PD is part of.}
\item{REGION_name}{Names of regions corresponding to the REGION ID.}
\item{geometry}{The sfc polygon geometry (boundaries).}
}
}
\source{
"2016 Transportation Tomorrow Survey" from the \href{http://dmg.utoronto.ca/survey-boundary-files}{Data Management Group} accessed February 28th 2022.
}
\usage{
data(ggh_pd)
}
\description{
This object contains the planning boundaries (PB) associated with the 2016 Transportation Tomorrow Survey (TTS) in the Greater Golden Horseshoe (GGH) area.
}
\keyword{boundary}
\keyword{districts}
\keyword{municipalities}
\keyword{planning}
\keyword{regions}
