% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TTS2016R-package.R
\docType{data}
\name{ggh_taz}
\alias{ggh_taz}
\title{Traffic analysis zones boundaries in the Greater Golden Horseshoe (GGH) area and associated data.}
\format{
A simple feature class (sf) polygon object containing 3764 rows and 10 variables; each row represents a unique TAZ with associated features.
\describe{
\item{GTA06}{Unique ID of the traffic analysis zone (TAZ).}
\item{AREA}{Area of TAZ in units of km^2.}
\item{REGION}{Retrieved from the ggh_pd object. It is a unique ID corresponding with the greater region that each planning district (PD) is part of so filtering and analysis, by PD, is simplified.}
\item{CMAUID}{Retrieved from the ggh_cma object. It is a unique identifier for each census metropolitan area (CMA)/census agglomeration (CA). Within, filtering and analysis that can be easily linked to the census data, is simplified.}
\item{workers}{The number of full-time workers in the TAZ.}
\item{jobs}{The number of full-time workers who declare their typical place of employment in the TAZ. }
\item{LTGO_split}{The percentage of full-time work related trips made from that origin to that destination using local transit and/or GO transit (regional transit) as a primary mode.}
\item{W_split}{The percentage of full-time work related trips made from that origin to that destination using a bicyle as a primary mode.}
\item{BC_split}{The percentage of full-time work related trips made from that origin to that destination by walking as a primary mode.}
\item{geometry}{The sfc polygon geometry (boundaries) of the TAZ.}
}
}
\source{
"2016 Transportation Tomorrow Survey" from \href{http://dmg.utoronto.ca/transportation-tomorrow-survey/tts-introduction}{Data Management Group} accessed November 14 2021.
}
\usage{
data(ggh_taz)
}
\description{
This object contains traffic analysis zones (TAZ) sourced from the 2016 Transportation Tomorrow Survey (TTS) in the Greater Golden Horseshoe (GGH) area.
}
\keyword{(GGH)}
\keyword{2016}
\keyword{Golden}
\keyword{Greater}
\keyword{Horseshoe}
\keyword{Jobs}
\keyword{TTS}
\keyword{Workers}
\keyword{area.}
\keyword{in}
\keyword{the}
