% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TTS2016R-package.R
\docType{data}
\name{od}
\alias{od}
\title{An origin and destination table of workers, place of employment, trip counts, and estimated travel times.}
\format{
A dataframe containing 103076 rows and 17 variables; each row represents a unique trip from origin TAZ to destination TAZ with associated features.
\describe{
\item{Origin}{The unique ID of the origin traffic analysis zone (TAZ). It is equivalent to the IDs used in 'GTA06'.}
\item{Destination}{The unique ID of the destination traffic analysis zone (TAZ). It is equivalent to the IDs used in 'GTA06'.}
\item{Persons}{The number of full-time workers at the origin who work at the destination.}
\item{T_LT}{The number of work-related trips taken from the origin to the destination by local transit excluding GO transit (regional transit service) as a primary mode.}
\item{T_LTGO}{The number of work-related trips taken from the origin to the destination by local transit and GO transit (regional transit service) as a primary mode.}
\item{T_GO}{The number of work-related trips taken from the origin to the destination by GO transit (regional transit service as a primary mode.}
\item{BC}{The number of work-related trips taken from the origin to the destination by bicyle (primary mode).}
\item{W}{The number of work-related trips taken from the origin to the destination by walking (primary mode).}
\item{C_DPC}{The number of work-related trips taken from the origin to the destination as a driver of a private car (primary mode).}
\item{C_PPC}{The number of work-related trips taken from the origin to the destination as a passenger of a private car (primary mode).}
\item{C_TA}{The number of work-related trips taken from the origin to the destination by a taxi (primary mode).}
\item{C_RS}{The number of work-related trips taken from the origin to the destination by a paid ride-share  (primary mode).}
\item{C_SB}{The number of work-related trips taken from the origin to the destination by a school bus (primary mode).}
\item{M}{The number of work-related trips taken from the origin to the destination by a motorcycle (primary mode).}
\item{O}{The number of work-related trips taken from the origin to the destination by another mode (primary mode).}
\item{sum_trips}{The total number of work-related trips taken from the origin to the destination by all modes.}
\item{travel_time}{The car travel time from origin to destination assuming a 7:00 am departure from and to the TAZ centroids on Oct. 20 2021.}
}
}
\source{
"2016 Transportation Tomorrow Survey" from the \href{http://dmg.utoronto.ca/transportation-tomorrow-survey/tts-introduction}{Data Management Group} accessed November 14 2021.

Travel times calculated using \href{https://github.com/ipeaGIT/r5r}{\code{r5r}}
}
\usage{
data(od)
}
\description{
This object contains the trips made from origin (TAZ) to destination (TAZ) by full-time workers to work destinations; it is sourced from the 2016 Transportation Tomorrow Survey (TTS) in the Greater Golden Horseshoe (GGH) area. Also contained within is the calculated car travel time for each trip calculated using TAZ centroids and r5r in October 2021.
}
\keyword{(GGH)}
\keyword{2016}
\keyword{Destination}
\keyword{Golden}
\keyword{Greater}
\keyword{Horseshoe}
\keyword{Jobs}
\keyword{Origin}
\keyword{TTS}
\keyword{Trips}
\keyword{Workers}
\keyword{area.}
\keyword{in}
\keyword{the}
