% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TTS2016R-package.R
\docType{data}
\name{od_ggh4_mode_purpose}
\alias{od_ggh4_mode_purpose}
\title{An origin and destination table using version GGHM V4 of the traffic analysis zones with trips by start_time, trip purpose of destination, and primary mode of travel.}
\format{
A dataframe containing 782364 rows and 7 variables; each row represents a origin-destination pair.
\describe{
\item{Origin}{The unique ID of the origin traffic analysis zone (TAZ). It is equivalent to the IDs used in 'GGHV4' in table 'ggh_taz_v4'.}
\item{Destination}{The unique ID of the destination traffic analysis zone (TAZ). It is equivalent to the IDs used in 'GGHV4' in table 'ggh_taz_v4'.}
\item{Trips}{The number of trips between the origin and destination.}
\item{start_time}{Start time of the trips; it is an object of class "period".}
\item{day}{Day of the week when the trips started; it is an ordered factor with days of the week.}
\item{purpose}{The purpose of the trips at the destination; it is a factor.}
\item{mode}{The main mode of travel for the trips; it is a factor.}
}
}
\source{
"2016 Transportation Tomorrow Survey" from the \href{http://dmg.utoronto.ca/transportation-tomorrow-survey/tts-introduction}{Data Management Group} accessed November 14 2021.
}
\usage{
data(od_ggh4_mode_purpose)
}
\description{
This object contains the trips recorded from origin (TAZ) to destination (TAZ) by time of day, day, purpose, and main mode of travel; it is sourced from the 2016 Transportation Tomorrow Survey (TTS) in the Greater Golden Horseshoe (GGH) area.
}
\details{
TTS collects trips only on weekdays (i.e., Monday to Friday), but some trips are recorded with start times after midnight. Such trips were assigned to the following day.

Trip purposes at destination:
\describe{
\item{Work}{First work trip of the day.}
\item{Subsequent work}{Second and subsequent work trips.}
\item{School}{First school trip of the day.}
\item{Subsequent school}{Second and subsequent school trips.}
\item{Daycare}{Daycare.}
\item{Home}{Home.}
\item{Facilitate}{Facilitate.}
\item{Market}{Marketing/Shopping.}
\item{Other}{Catch all category for all trips with purposes distinct to those above.}
}
Primary mode of the trip (if any part of the trip is made by public transit, the primary mode is defined as B, G, or J):
\describe{
\item{Transit}{Public transit excluding GO rail.}
\item{Cycle}{Bicycle.}
\item{Driver}{Auto driver.}
\item{GO}{GO rail.}
\item{GO+Transit}{Joint GO rail and local public transit.}
\item{Motorcycle}{Motorcycle.}
\item{Other}{Other modes.}
\item{Passenger}{Auto passenger.}
\item{School}{School bus.}
\item{Taxi}{Taxi passenger.}
\item{Rideshare}{Paid rideshare.}
\item{Walk}{Walk trips are only collected for trips to and from work or school or when the walk trip is a connecting trip within a chain of trips that use other non-walk modes.}
}

NOTE: TTS collects trips taken for unknown purposes and/or using an unknown mode, but these were not downloaded for this data package. There is a small number of trips ($n = 7,368$) that where taken for an unknown purpose by any mode and on any day. There are zero trips that where undertaken by an unkown mode for any purpose on any day.
In other words, this table is incomplete, but the number of trips that are missing is very small.
}
\keyword{(GGH)}
\keyword{2016}
\keyword{Destination}
\keyword{Golden}
\keyword{Greater}
\keyword{Horseshoe}
\keyword{Origin}
\keyword{TTS}
\keyword{Trips}
\keyword{area.}
\keyword{in}
\keyword{mode}
\keyword{purpose}
\keyword{the}
