% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TTS2016R-package.R
\docType{data}
\name{toy_od_table}
\alias{toy_od_table}
\title{toy_od_table}
\format{
A dataframe containing 27 rows and 7 variables; each row represents a unique trip from "Population" to "Employment Center" with associated features.
\describe{
\item{Origin}{The unique ID for each "Population" from 1 to 9 }
\item{Destination}{The unique ID for each "Employment Center" from 1 to 3}
\item{Population}{The population corresponding to the "Population" ID.}
\item{Jobs}{The number of jobs corresponding to the "Employment Center" ID.}
\item{distance}{The distance between origin to destination (unitless).}
\item{catchments}{A binary code indicating if the origin-destination trip is eligible or not (details on how this can be used within the vignette).}
\item{trips}{Randomly generated number of trips made from origin to destination.}
}
}
\usage{
data(toy_od_table)
}
\description{
This object contains a hypothetical toy example of Population, Jobs, Distance, and Catchment for each Origin-Destination trip. Randomly generated.
}
\keyword{Destination}
\keyword{Jobs}
\keyword{Origin}
\keyword{Toy}
\keyword{Trips}
\keyword{Workers}
