% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TTS2016R-package.R
\docType{data}
\name{toy_sim_zones}
\alias{toy_sim_zones}
\title{toy_sim_zones}
\format{
A simple feature class (sfc) multipolygon object containing 12 rows and 5 variables; each row represents a unique "Population" or "Employment Center" zone with associated features.
\describe{
\item{id}{The unique ID for each "Population" (from 1 to 9) or "Employment Center" (from 1 to 3) }
\item{id_short}{A short unique ID}
\item{number}{The population corresponding to the ID.}
\item{type}{The number of jobs corresponding the ID.}
\item{geometry}{The sfc polygon geometry (boundaries) of each ID.}
}
}
\usage{
data(toy_sim_zones)
}
\description{
This object contains the geometries of the hypothetical toy example Origin and Destination trip zones. Randomly generated.
}
\keyword{Destination}
\keyword{Jobs}
\keyword{Origin}
\keyword{Toy}
\keyword{Workers}
