% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TTS2016R-package.R
\docType{package}
\name{TTS2016R-package}
\alias{TTS2016R}
\alias{TTS2016R-package}
\title{TTS2016R: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\references{
\url{https://github.com/soukhova/TTS2016R}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/soukhova/TTS2016R}
  \item \url{https://soukhova.github.io/TTS2016R/}
  \item Report bugs at \url{https://github.com/soukhova/TTS2016R/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})


Antonio Paez, School of Earth, Environment and Society, McMaster University \email{paezha@mcmaster.ca}

Anastasia Soukhov, School of Earth, Environment and Society, McMaster University \email{soukhoa@mcmaster.ca}
}
\keyword{internal}
