% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TTS2016R-package.R
\docType{package}
\name{TTS2016R}
\alias{TTS2016R}
\title{TTS2016R}
\description{
A data package which consists of objects sourced from the
\href{http://dmg.utoronto.ca/transportation-tomorrow-survey/tts-introduction}{2016 Transportation Tomorrow Survey (TTS)}
or are curated to facilitate the use and analysis of TTS data. This package includes worker-to-jobs origin
-destinations and traffic analysis zone (TAZ) boundaries for the Greater Golden Horse area (GGH) located in
southern Ontario, Canada. The package also includes associated spatial boundary files.
}
\references{
\url{https://github.com/soukhova/TTS2016R}
}
\author{
Antonio Paez, School of Earth, Environment and Society, McMaster University \email{paezha@mcmaster.ca}

Anastasia Soukhov, School of Earth, Environment and Society, McMaster University \email{soukhoa@mcmaster.ca}
}
