% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TTS2016R-package.R
\docType{data}
\name{ggh_taz}
\alias{ggh_taz}
\title{ggh_taz}
\format{
A simple feature class (sf) polygon object containing 3764 rows and 5 variables; each row represents a unique TAZ with associated features.
\describe{
\item{GTA06}{Unique ID of the traffic analysis zone (TAZ).}
\item{AREA}{Area of TAZ in units of km^2.}
\item{workers}{The number of full-time workers within each TAZ.}
\item{jobs}{The number of jobs within each TAZd. }
\item{geometry}{The sfc polygon geometry (boundaries) of the TAZ.}
}
}
\source{
"2016 Transportation Tomorrow Survey" from \href{http://dmg.utoronto.ca/transportation-tomorrow-survey/tts-introduction}{Data Management Group} accessed November 14 2021.
}
\usage{
data(ggh_taz)
}
\description{
This object contains traffic analysis zones (TAZ) sourced from the 2016 Transportation Tomorrow Survey (TTS) in the Greater Golden Horseshoe (GGH) area.
}
\keyword{(GGH)}
\keyword{2016}
\keyword{Golden}
\keyword{Greater}
\keyword{Horseshoe}
\keyword{Jobs}
\keyword{TTS}
\keyword{Workers}
\keyword{area.}
\keyword{in}
\keyword{the}
