% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TTS2016R-package.R
\docType{data}
\name{ggh_taz}
\alias{ggh_taz}
\alias{ggh_taz_v4}
\alias{hh_sed_taz}
\title{Traffic analysis zones boundaries in the Greater Golden Horseshoe (GGH) area and associated data.}
\format{
A simple feature class (sf) polygon object containing 3764 rows and 10 variables; each row represents a unique TAZ with associated features.
\describe{
\item{GTA06}{Unique ID of the traffic analysis zone (TAZ).}
\item{AREA}{Area of TAZ in units of km^2.}
\item{REGION}{Retrieved from the ggh_pd object. It is a unique ID corresponding with the greater region that each planning district (PD) is part of so filtering and analysis, by PD, is simplified.}
\item{CMAUID}{Retrieved from the ggh_cma object. It is a unique identifier for each census metropolitan area (CMA)/census agglomeration (CA). Within, filtering and analysis that can be easily linked to the census data, is simplified.}
\item{workers}{The number of full-time workers in the TAZ.}
\item{jobs}{The number of full-time workers who declare their typical place of employment in the TAZ. }
\item{LTGO_split}{The percentage of full-time work related trips made from that origin to that destination using local transit and/or GO transit (regional transit) as a primary mode.}
\item{W_split}{The percentage of full-time work related trips made from that origin to that destination using a bicyle as a primary mode.}
\item{BC_split}{The percentage of full-time work related trips made from that origin to that destination by walking as a primary mode.}
\item{geometry}{The sfc polygon geometry (boundaries) of the TAZ.}
}

A simple feature class (sf) polygon object containing 3258 rows and 7 variables; each row represents a unique TAZ with associated features. The object is projected as WGS 84 / UTM zone 17N (epsg: 32617).
\describe{
\item{GGHV4}{Unique ID of the traffic analysis zone (TAZ).}
\item{CSDUID}{Unique ID corresponding with the greater region that each Census Subdivision (CSD) is part of so filtering and analysis, by CSD, is simplified.}
\item{CSDNAME}{Name of the Census Subdivision (CSD).}
\item{CMAUID}{Unique identifier for each census metropolitan area (CMA)/census agglomeration (CA). Within, filtering and analysis that can be easily linked to the census data, is simplified.}
\item{CMAUID}{Retrieved from the ggh_cma object. It is a unique identifier for each census metropolitan area (CMA)/census agglomeration (CA). Within, filtering and analysis that can be easily linked to the census data, is simplified.}
\item{CMANAME}{Name of the census metropolitan area (CMA)/census agglomeration (CA).}
\item{workers}{The number of full-time workers in the TAZ.}
\item{jobs}{The number of full-time workers who declare their typical place of employment in the TAZ. }
\item{LTGO_split}{The percentage of full-time work related trips made from that origin to that destination using local transit and/or GO transit (regional transit) as a primary mode.}
\item{W_split}{The percentage of full-time work related trips made from that origin to that destination using a bicyle as a primary mode.}
\item{BC_split}{The percentage of full-time work related trips made from that origin to that destination by walking as a primary mode.}
\item{geometry}{The sfc polygon geometry (boundaries) of the TAZ.}
}

A data frame containing 3128 rows and 8 variables; each row represents a unique TAZ with associated features.
\describe{
\item{GTA06}{Unique ID of the traffic analysis zone (TAZ).}
\item{$0 to $14999}{Number of households in TAZ in this income range. }
\item{$15000 to $39999}{Number of households in TAZ in this income range.}
\item{$40000 to $59999}{Number of households in TAZ in this income range.}
\item{$60000 to $99999}{Number of households in TAZ in this income range.}
\item{$100000 to $124999}{Number of households in TAZ in this income range. }
\item{$125000 and above}{TNumber of households in TAZ in this income range.}
\item{Decline / do not know}{Number of households in TAZ who reportet unknown or did not state their income range.}
}
}
\source{
"2016 Transportation Tomorrow Survey" from \href{http://dmg.utoronto.ca/transportation-tomorrow-survey/tts-introduction}{Data Management Group} accessed November 14 2021.

"2016 Transportation Tomorrow Survey" from \href{http://dmg.utoronto.ca/transportation-tomorrow-survey/tts-introduction}{Data Management Group} accessed November 14 2021.

"2016 Transportation Tomorrow Survey" from \href{http://dmg.utoronto.ca/transportation-tomorrow-survey/tts-introduction}{Data Management Group} accessed November 14 2021.
}
\usage{
data(ggh_taz)

data(ggh_taz_v4)

data(ggh_taz)
}
\description{
This object contains traffic analysis zones (TAZ) sourced from the 2016 Transportation Tomorrow Survey (TTS) in the Greater Golden Horseshoe (GGH) area. The object is projected as WGS 84 / UTM zone 17N (epsg: 32617).

This object contains traffic analysis zones (TAZ) sourced from the 2016 Transportation Tomorrow Survey (TTS) in the Greater Golden Horseshoe (GGH) area.

This object contains socio-economic and demographic information at the level of traffic analysis zones (TAZ) sourced from the 2016 Transportation Tomorrow Survey (TTS) in the Greater Golden Horseshoe (GGH) area.
The query to retrieve data for income was a cross tabulation of the Person table; the Row was 2006 GTA zone of household - gta06_hhld, and the column was Income range of household - income.
The income reported is the total  for the household from all sources, before income tax, in the year before the survey.
According to the TTS Data Guide "The question was voluntary and subject to high levels of non-response. The results of this question should be interpreted with caution."
}
\keyword{(GGH)}
\keyword{2016}
\keyword{Golden}
\keyword{Greater}
\keyword{Horseshoe}
\keyword{Jobs}
\keyword{TTS}
\keyword{Workers}
\keyword{area.}
\keyword{in}
\keyword{the}
