% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TTS2016R-package.R
\docType{package}
\name{TTS2016R-package}
\alias{TTS2016R}
\alias{TTS2016R-package}
\title{TTS2016R: An augmented 2016 Transportation Tomorrow Survey (TTS) data package: trips and estimated travel time to work in the Greater Golden Horsehoe area, Canada}
\description{
This package includes person-to-jobs origin-destinations, the traffic analysis zone (TAZ) boundaries, and the associated municipality boundaries from the 2016 [TTS](http://dmg.utoronto.ca/transportation-tomorrow-survey/tts-introduction) of the Greater Golden Horse area (GGH) located in southern Ontario, Canada.
}
\references{
\url{https://github.com/soukhova/TTS2016R}
}
\author{
\strong{Maintainer}: Anastasia Soukhov \email{soukhoa@mcmaster.ca} (\href{https://orcid.org/0000-0003-4371-4831}{ORCID})

Authors:
\itemize{
  \item Antonio Paez \email{paezha@mcmaster.ca} (\href{https://orcid.org/0000-0001-6912-9919}{ORCID})
}


Antonio Paez, School of Earth, Environment and Society, McMaster University \email{paezha@mcmaster.ca}

Anastasia Soukhov, School of Earth, Environment and Society, McMaster University \email{soukhoa@mcmaster.ca}
}
\keyword{internal}
